+++ backend/api/newsletters/forms.py
@@ -1,10 +1,10 @@
 from django import forms
 from newsletters.models import Subscription
 from strawberry_forms.forms import FormWithContext
-from api.newsletters.mailchimp import subscribe
+from integrations.mailchimp import subscribe
 
 
-class SubscribeToNewsletterForm(FormWithContext):
+class SussbscribeToNewsletterForm(FormWithContext):
     email = forms.EmailField()
 
     def save(self):
